HW1/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.22.0
pandas>=1.4.0
matplotlib>=3.5.0
tqdm>=4.62.0
scikit-learn>=1.0.0
Pillow>=9.0.0
